+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed tappable items from Appointment Unit card [#3127](https://github.com/rokwire/illinois-app/issues/3127).
 - Open group post urls in an external browser [#3129](https://github.com/rokwire/illinois-app/issues/3129).
 - Load mental health buildings mapping from content service [#3131](https://github.com/rokwire/illinois-app/issues/3131).
+
 ## [4.3.41] - 2023-04-12
 ### Added
 - My College of Medicine Courses [#3118](https://github.com/rokwire/illinois-app/issues/3118).